fix: albums and tracks routes

+++ src/albums/albums.controller.ts
@@ -3,9 +3,13 @@ import {
   Get,
   Post,
   Body,
-  Patch,
   Param,
   Delete,
+  ParseUUIDPipe,
+  NotFoundException,
+  Put,
+  ForbiddenException,
+  HttpCode,
 } from '@nestjs/common';
 import { AlbumsService } from './albums.service';
 import { CreateAlbumDto } from './dto/create-album.dto';
@@ -18,7 +22,9 @@ export class AlbumsController {
 
   @Post()
   create(@Body() createAlbumDto: CreateAlbumDto) {
-    return this.albumsService.create(createAlbumDto);
+    const { name, year, artistId } = createAlbumDto ?? {};
+
+    return this.albumsService.create(new CreateAlbumDto(name, year, artistId));
   }
 
   @Get()
@@ -27,29 +33,41 @@ export class AlbumsController {
   }
 
   @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.albumsService.findOne(id);
-  }
-
-  @Patch(':id')
-  update(@Param('id') id: string, @Body() updateAlbumDto: UpdateAlbumDto) {
-    const album = this.findOne(id) as Album | undefined;
+  findOne(@Param('id', new ParseUUIDPipe()) id: string) {
+    const album = this.albumsService.findOne(id);
 
     if (!album) {
-      return 'error';
+      throw new NotFoundException();
     }
+
+    return album;
+  }
+
+  @Put(':id')
+  update(
+    @Param('id', new ParseUUIDPipe()) id: string,
+    @Body() updateAlbumDto: UpdateAlbumDto,
+  ) {
+    const album = this.findOne(id) as Album | undefined;
 
     const updatedAlbum = this.albumsService.update(album, updateAlbumDto);
 
     if (!updatedAlbum) {
-      return 'error';
+      throw new ForbiddenException();
     }
 
     return updatedAlbum;
   }
 
+  @HttpCode(204)
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  remove(@Param('id', new ParseUUIDPipe()) id: string) {
+    const isRemoved = this.albumsService.remove(id);
+
+    if (!isRemoved) {
+      throw new NotFoundException();
+    }
+
     return this.albumsService.remove(id);
   }
 }
